Update use of XDG Base Dir. to comply with spec.
According to the XDG Base Directory Specification [1],

"$XDG_CONFIG_HOME defines the base directory relative to which user
specific configuration files should be stored. If $XDG_CONFIG_HOME
is either not set or empty, a default equal to
$HOME/.config should be used. "

Not doing so means that dmenfm won't work quite right, if your
system doesn't set `$XDG_CONFIG_HOME`. Because a default is to be
assumed, many systems do not set this variable, so the default
really must be included in the script.

[1] http://standards.freedesktop.org/basedir-spec/latest/ar01s03.html